requirements.txt
google-generativeai
langchain
langchain-core
langchain-text-splitters
sentence-transformers
pypdf
faiss-cpu
nltk
pandas
numpy
scikit-learn
ipywidgets
tqdm